Media functions to work with group image
